Added `spree g authentication:devise`
Rather than using `spree_auth_devise` we will use a lightweight generator to setup Spree with Devise. In the future we will also add Rails 8 default authentication as an option.
