--- core/lib/generators/spree/custom_user/templates/current_user_helpers.rb.tt
@@ -1,15 +1,0 @@
-module Spree
-  module CurrentUserHelpers
-    def self.included(receiver)
-      receiver.send :helper_method, :spree_current_user
-    end
-
-    def spree_current_user
-      current_user
-    end
-  end
-end
-
-
-ApplicationController.include Spree::CurrentUserHelpers if defined?(ApplicationController)
-Spree::Api::BaseController.include Spree::CurrentUserHelpers if defined?(Spree::Api::BaseController)
