--- core/lib/generators/spree/custom_user/templates/initializer.rb.tt
@@ -1,1 +1,0 @@
-Spree.user_class = "<%= class_name %>"
