--- core/lib/generators/spree/custom_user/templates/migration.rb.tt
@@ -1,7 +1,0 @@
-class AddSpreeFieldsToCustomUserTable < ActiveRecord::Migration[4.2]
-  def up
-    add_column <%= table_name.inspect %>, :spree_api_key, :string, limit: 48
-    add_column <%= table_name.inspect %>, :ship_address_id, :bigint
-    add_column <%= table_name.inspect %>, :bill_address_id, :bigint
-  end
-end
