+++ storefront/app/views/devise/passwords/edit.html.erb
@@ -18,5 +18,5 @@
       <%= f.submit Spree.t(:reset_password), class: 'btn btn-primary' %>
     </div>
   <% end %>
-  <%= render "spree/shared/devise/links" %>
+  <%= render "devise/shared/links" %>
 </div>
