+++ storefront/app/views/devise/passwords/new.html.erb
@@ -11,6 +11,6 @@
         <%= f.submit Spree.t(:reset_password), class: 'btn btn-primary w-full' %>
       </div>
     <% end %>
-    <%= render "spree/shared/devise/links" %>
+    <%= render "devise/shared/links" %>
   </div>
 <% end %>
