+++ storefront/app/views/devise/registrations/_form.html.erb
