+++ storefront/app/views/devise/registrations/new.html.erb
@@ -22,6 +22,6 @@
         <%= f.submit Spree.t(:sign_up), class: 'btn btn-primary w-full' %>
       </div>
     <% end %>
-    <%= render "spree/shared/devise/links" %>
+    <%= render "devise/shared/links" %>
   </div>
 <% end%>
