+++ storefront/app/views/devise/sessions/new.html.erb
@@ -1,7 +1,7 @@
 <%= turbo_frame_tag :login do %>
   <div class="card-dialog max-w-xl mx-auto py-10 px-4">
     <h2 class="text-xl lg:text-2xl font-medium mb-4"><%= Spree.t(:login) %></h2>
-    <%= form_for(resource, as: resource_name, url: session_path(resource_name), data: { turbo: false }) do |f| %>
+    <%= form_for(resource, as: resource_name, url: spree_login_path, data: { turbo: false }) do |f| %>
       <div class="flex flex-col mb-4">
         <%= f.label :email, Spree.t(:email) %>
         <%= f.email_field :email, autofocus: true, autocomplete: "email", class: 'text-input', required: true %>
@@ -20,6 +20,6 @@
         <%= f.submit Spree.t(:login), class: 'btn btn-primary w-full' %>
       </div>
     <% end %>
-    <%= render "spree/shared/devise/links" %>
+    <%= render "devise/shared/links" %>
   </div>
 <% end %>
